+++ CopilotKit/packages/backend/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @copilotkit/cloud
 
+## 0.4.0-mme-fix-tools-error.2
+
+### Minor Changes
+
+- Don't send empty tools array to OpenAI
+
+## 0.4.0-beta-automation.1
+
+### Minor Changes
+
+- Test automation
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.4.0-beta-automation.1
+
 ## 0.4.0-tools.0
 
 ### Minor Changes